Run Firefox acceptance inside project-file CI gate
